City

+++ src/main/java/airline/data/CountryJpaController.java
@@ -1,13 +1,16 @@
 package airline.data;
 
-import airline.exceptions.NonexistentEntityException;
-import airline.exceptions.PreexistingEntityException;
-import airline.logic.Country;
+import airline.exceptions.*;
 import java.io.Serializable;
-import java.util.List;
-import javax.persistence.*;
+import javax.persistence.Query;
+import javax.persistence.EntityNotFoundException;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
+import airline.logic.City;
+import airline.logic.Country;
+import java.util.ArrayList;
+import java.util.List;
+import javax.persistence.*;
 
 public class CountryJpaController implements Serializable
 {
@@ -24,12 +27,34 @@ public class CountryJpaController implements Serializable
 
   public void create(Country country) throws PreexistingEntityException, Exception
   {
+    if (country.getCityList() == null)
+    {
+      country.setCityList(new ArrayList<City>());
+    }
     EntityManager em = null;
     try
     {
       em = getEntityManager();
       em.getTransaction().begin();
+      List<City> attachedCityList = new ArrayList<City>();
+      for (City cityListCityToAttach : country.getCityList())
+      {
+        cityListCityToAttach = em.getReference(cityListCityToAttach.getClass(), cityListCityToAttach.getCode());
+        attachedCityList.add(cityListCityToAttach);
+      }
+      country.setCityList(attachedCityList);
       em.persist(country);
+      for (City cityListCity : country.getCityList())
+      {
+        Country oldCountryOfCityListCity = cityListCity.getCountry();
+        cityListCity.setCountry(country);
+        cityListCity = em.merge(cityListCity);
+        if (oldCountryOfCityListCity != null)
+        {
+          oldCountryOfCityListCity.getCityList().remove(cityListCity);
+          oldCountryOfCityListCity = em.merge(oldCountryOfCityListCity);
+        }
+      }
       em.getTransaction().commit();
     }
     catch (Exception ex)
@@ -49,14 +74,55 @@ public class CountryJpaController implements Serializable
     }
   }
 
-  public void edit(Country country) throws NonexistentEntityException, Exception
+  public void edit(Country country) throws IllegalOrphanException, NonexistentEntityException, Exception
   {
     EntityManager em = null;
     try
     {
       em = getEntityManager();
       em.getTransaction().begin();
+      Country persistentCountry = em.find(Country.class, country.getCode());
+      List<City> cityListOld = persistentCountry.getCityList();
+      List<City> cityListNew = country.getCityList();
+      List<String> illegalOrphanMessages = null;
+      for (City cityListOldCity : cityListOld)
+      {
+        if (!cityListNew.contains(cityListOldCity))
+        {
+          if (illegalOrphanMessages == null)
+          {
+            illegalOrphanMessages = new ArrayList<String>();
+          }
+          illegalOrphanMessages.add("You must retain City " + cityListOldCity + " since its country field is not nullable.");
+        }
+      }
+      if (illegalOrphanMessages != null)
+      {
+        throw new IllegalOrphanException(illegalOrphanMessages);
+      }
+      List<City> attachedCityListNew = new ArrayList<City>();
+      for (City cityListNewCityToAttach : cityListNew)
+      {
+        cityListNewCityToAttach = em.getReference(cityListNewCityToAttach.getClass(), cityListNewCityToAttach.getCode());
+        attachedCityListNew.add(cityListNewCityToAttach);
+      }
+      cityListNew = attachedCityListNew;
+      country.setCityList(cityListNew);
       country = em.merge(country);
+      for (City cityListNewCity : cityListNew)
+      {
+        if (!cityListOld.contains(cityListNewCity))
+        {
+          Country oldCountryOfCityListNewCity = cityListNewCity.getCountry();
+          cityListNewCity.setCountry(country);
+          cityListNewCity = em.merge(cityListNewCity);
+          if (oldCountryOfCityListNewCity != null && !oldCountryOfCityListNewCity.equals(country))
+          {
+            oldCountryOfCityListNewCity.getCityList().remove(cityListNewCity);
+            oldCountryOfCityListNewCity = em.merge(oldCountryOfCityListNewCity);
+          }
+        }
+      }
       em.getTransaction().commit();
     }
     catch (Exception ex)
@@ -81,7 +147,7 @@ public class CountryJpaController implements Serializable
     }
   }
 
-  public void destroy(String id) throws NonexistentEntityException
+  public void destroy(String id) throws IllegalOrphanException, NonexistentEntityException
   {
     EntityManager em = null;
     try
@@ -97,6 +163,21 @@ public class CountryJpaController implements Serializable
       catch (EntityNotFoundException enfe)
       {
         throw new NonexistentEntityException("The country with id " + id + " no longer exists.", enfe);
+      }
+      List<String> illegalOrphanMessages = null;
+      List<City> cityListOrphanCheck = country.getCityList();
+      for (City cityListOrphanCheckCity : cityListOrphanCheck)
+      {
+        if (illegalOrphanMessages == null)
+        {
+          illegalOrphanMessages = new ArrayList<String>();
+        }
+        illegalOrphanMessages.add("This Country (" + country + ") cannot be destroyed since the City "
+                + cityListOrphanCheckCity + " in its cityList field has a non-nullable country field.");
+      }
+      if (illegalOrphanMessages != null)
+      {
+        throw new IllegalOrphanException(illegalOrphanMessages);
       }
       em.remove(country);
       em.getTransaction().commit();
